src/translation/event_listener/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-logging
google-api-python-client
google-auth
jsonpath-ng